fix: fix newLine
